Reuse load and write json functions
