+++ scripts/migration-project_15_0_0/migration/submit_projects.py
@@ -1,0 +1,3 @@
+if __name__ == '__main__':
+    submi
+    update_projects(projects)
