+++ scripts/migration-project_15_0_0/perf/format_xls.py
@@ -1,11 +1,6 @@
 import json
 
-
-def load_json(filename: str):
-    with open(filename) as json_file:
-        data = json.load(json_file)
-        return data
-
+from util import load_json
 
 if __name__ == '__main__':
     submissions_by_project = load_json('summary.json')
